refactor: move error to its own module
