+++ apollo-router/src/plugins/response_cache/storage/mod.rs
@@ -1,102 +1,34 @@
-pub(crate) mod postgres;
+pub(super) mod error;
+pub(super) mod postgres;
 
 use std::collections::HashMap;
-use std::fmt::Display;
-use std::fmt::Formatter;
 use std::time::Duration;
 use std::time::Instant;
 
-use serde_json::error::Category;
 use tokio_util::future::FutureExt;
 
-use crate::plugins::response_cache::ErrorCode;
-use crate::plugins::response_cache::cache_control::CacheControl;
+use super::cache_control::CacheControl;
 
-#[derive(Debug)]
-pub(crate) enum Error {
-    Database(sqlx::Error),
-    Serialize(serde_json::Error),
-    Timeout,
-}
-
-impl Error {
-    pub(super) fn is_row_not_found(&self) -> bool {
-        match self {
-            Error::Database(err) => matches!(err, &sqlx::Error::RowNotFound),
-            Error::Serialize(_) => false,
-            Error::Timeout => false,
-        }
-    }
-}
-
-impl Display for Error {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Error::Database(err) => f.write_str(&err.to_string()),
-            Error::Serialize(err) => f.write_str(&err.to_string()),
-            Error::Timeout => f.write_str("TIMED_OUT"),
-        }
-    }
-}
-
-impl From<sqlx::Error> for Error {
-    fn from(err: sqlx::Error) -> Self {
-        Error::Database(err)
-    }
-}
-
-impl From<serde_json::Error> for Error {
-    fn from(err: serde_json::Error) -> Self {
-        Error::Serialize(err)
-    }
-}
-
-impl From<tokio::time::error::Elapsed> for Error {
-    fn from(_: tokio::time::error::Elapsed) -> Self {
-        Error::Timeout
-    }
-}
-
-impl ErrorCode for Error {
-    fn code(&self) -> &'static str {
-        match self {
-            Error::Database(err) => err.code(),
-            Error::Serialize(err) => match err.classify() {
-                Category::Io => "Serialize::IO",
-                Category::Syntax => "Serialize::Syntax",
-                Category::Data => "Serialize::Data",
-                Category::Eof => "Serialize::EOF",
-            },
-            Error::Timeout => "TIMED_OUT",
-        }
-    }
-}
-
-impl std::error::Error for Error {}
-
-type StorageResult<T> = Result<T, Error>;
+type StorageResult<T> = Result<T, self::error::Error>;
 
 type Documents = Vec<Document>;
 
 #[derive(Debug, Clone)]
-pub(crate) struct Document {
-    pub(crate) cache_key: String,
-    pub(crate) data: serde_json_bytes::Value,
-    pub(crate) cache_control: CacheControl,
-    pub(crate) invalidation_keys: Vec<String>,
-    pub(crate) expire: Duration,
+pub(super) struct Document {
+    pub(super) cache_key: String,
+    pub(super) data: serde_json_bytes::Value,
+    pub(super) cache_control: CacheControl,
+    pub(super) invalidation_keys: Vec<String>,
+    pub(super) expire: Duration,
 }
 
 #[derive(Debug, Clone)]
-pub(crate) struct CacheEntry {
-    pub(crate) cache_key: String,
-    pub(crate) data: serde_json_bytes::Value,
-    pub(crate) control: CacheControl,
+pub(super) struct CacheEntry {
+    pub(super) cache_key: String,
+    pub(super) data: serde_json_bytes::Value,
+    pub(super) control: CacheControl,
 }
 
-// TODO: in theory, we could have `struct Storage<S: CacheStorage>`. But the types are a huge pain.
-//  Keeping this trait around for now as it provides clear expected cache behavior, but not sure if
-//  that's actually good practice.
 pub(super) trait CacheStorage {
     fn timeout_duration(&self) -> Duration;
